Move collector from main to separate package
